extracts FakeHumanPlayer into separate file
